+++ stroom-ui/src/components/Account/TokenManager/TokenListDialog.stories.tsx
@@ -1,20 +1,20 @@
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
-import { AccountListDialog } from "./AccountListDialog";
+import { TokenListDialog } from "./TokenListDialog";
 
 import fullTestData from "testing/data";
-import useColumns from "./useColumns";
+import useTokenManager from "./useTokenManager";
 
-const stories = storiesOf("Account", module);
+const stories = storiesOf("Token", module);
 stories.add("List", () => {
-  const columns = useColumns();
+  const { columns } = useTokenManager();
 
   return (
-    <AccountListDialog
+    <TokenListDialog
       itemManagerProps={{
         tableProps: {
           columns: columns,
-          data: fullTestData.users,
+          data: fullTestData.tokens,
         },
         actions: {
           onCreate: () => undefined,